+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### Ignition Rendering 3.0.0 (20XX-XX-XX)
 
+1. Added function to get the number of channels from a GpuRay.
+    * [Pull request 178](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/178)
+
 ### Ignition Rendering 2.X
 
 ### Ignition Rendering 2.0.1